% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_allocateCellsForest.R
\name{allocateCellsForest}
\alias{allocateCellsForest}
\title{Allocate cells based on trained randomForest.}
\usage{
allocateCellsForest(newData, RForestObj, genesTrained)
}
\arguments{
\item{newData}{Dataframe or matrix containing the counts (genes x cells).
These should be from the cells you want to predict the cluster of.}

\item{RForestObj}{randomForest object already trained. To train a dataset, you
can use \code{\link[anSeq]{trainForest}}.}

\item{genesTrained}{vector of the genes that were used to train
the randomForest.}
}
\value{
dataframe with predicted allocations and probabilities.
}
\description{
For each cell, it allocates it to its most likely cluster
it belongs to from the trained random Forest.
}
\references{
\code{\link[randomForest]{randomForest}}.
}
\seealso{
\code{\link[randomForest]{randomForest}},
\code{\link[anSeq]{trainForest}}.
}
\author{
Blanca Pijuan Sala.
}
