% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_findMarkers.R
\name{findMarkers}
\alias{findMarkers}
\title{Find markers within groups}
\usage{
findMarkers(counts, genes = NULL, groups, topGeneThres = 1000,
  logFCThres = 0, qvalue = 0.05, mode = "general",
  groupTarget = NULL)
}
\arguments{
\item{counts}{Dataframe or matrix containing the counts (genes x cells).}

\item{genes}{vector of genes to use for analysis. Default: rownames(counts). It is important that
this list has been processed (i.e. not containing genes which all counts are 0). You can set this 
to the highly variable genes of the dataset. You can compute them using \code{findHVG}.}

\item{groups}{Vector specifying what group each cell belongs to. e.g. 
c("group1","group2","group1"...). This vector should be labelled (names()) by the cell name.}

\item{topGeneThres}{Threshold of upregulated genes to consider. 
Only used for 'strict' option. Default: 50.}

\item{logFCThres}{Threshold of log Fold Change to consider. 
Only used for 'strict' option. Default = 0.}

\item{qvalue}{FDR q-value to consider for threshold when testing. Default: 0.05.}

\item{mode}{"strict" or "general". "strict" will look for upregulated
genes by (1) doing pairwise comparisons between each of the groups and then (2) only
giving those genes that are upregulated in all the comparisons from group A
to all others e.g. if gene 1 is upregulated only when group A is compared to B
but not when A is compared to C, then it will not consider it as a marker gene. Instead, 
if the total number of groups is A, B, C (3), and if gene 1 is upregulated in A when
 comparing  A to B and A to C, then you will retrieve it."general" will make pairwise
 comparisons of group A to all the others as a group and output both upregulated and
 downregulated genes in a list of matrices. Default: "general".}

\item{groupTarget}{Set this to one label to perform DE between that group
i.e. group target, vs the rest of the cells.}
}
\value{
list of markers for each group.
}
\description{
It finds markers of particular groups within the dataset by comparing them to 
the other groups.
}
\author{
Blanca Pijuan Sala.
}
