% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_findSurfaceMarkers.R
\name{findSurfaceMarkers}
\alias{findSurfaceMarkers}
\title{Find markers within groups}
\usage{
findSurfaceMarkers(data, Reference = MouseSurface)
}
\arguments{
\item{data}{vector of gene Names.}

\item{Reference}{Vector with names being the ENTREZ gene symbols and elements being the 
confidence of being a surface protein. Default: MouseSurface. You can set it to human by doing
HumanSurface}
}
\description{
It finds markers of particular groups within the dataset by comparing them to 
the other groups.
}
\author{
Blanca Pijuan-Sala.
}
