% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getGeneIDs_getGeneNames.R
\name{getGeneID}
\alias{getGeneID}
\title{Obtain geneIDs from geneNames}
\usage{
getGeneID(geneNames, Table = geneTable)
}
\arguments{
\item{geneNames}{vector.}

\item{Table}{By default, it uses mm10 geneID - geneName association from Ensembl.
The table should have the columns named as 'Gene.ID' and 'Associated.Gene.Name'}
}
\value{
List with table of associations and vector of geneIDs.
}
\description{
It takes a vector of geneNames and looks for the geneIDs.
}
\author{
Blanca Pijuan-Sala.
}
