% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getGeneIDs_getGeneNames.R
\name{getGeneName}
\alias{getGeneName}
\title{Obtain geneNames from geneIDs}
\usage{
getGeneName(geneIDs, Table = geneTable)
}
\arguments{
\item{geneIDs}{vector.}

\item{Table}{By default, it uses mm10 geneID - geneName association from Ensembl.
The table should have the columns named as 'Gene.ID' and 'Associated.Gene.Name'}
}
\value{
List with table of associations and vector of geneNames.
}
\description{
It takes a vector of geneIDs and looks for the geneNames
}
\author{
Blanca Pijuan-Sala.
}
