% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_hclustCells.R
\name{hclustCells}
\alias{hclustCells}
\title{Cluster cells}
\usage{
hclustCells(data, clust.method = "ward.D2", leafreorder = "column")
}
\arguments{
\item{data}{Dataframe or matrix containing the counts (genes x cells). The 
algorithm will cluster on these.}

\item{clust.method}{Clustering method to use with hclust. Default: WarD2.
For more options, see \code{\link[stats]{hclust}}.}

\item{leafreorder}{Specify whether you want to obtain an hclust object
with reordered gene ("row") or cell leaves ("column"). If you want
to obtain both as a list, specify "both". Default: column (it will 
reorder the cells).}
}
\value{
hclust object
}
\description{
Cluster cells using the spearman rank correlation
as a similarity metric and (1-correlation)/2 as a distance metric.
}
\references{
\code{\link[stats]{hclust}}, \code{\link[cba]{order.optimal}}
}
\seealso{
\code{\link[stats]{hclust}}.
}
\author{
Blanca Pijuan Sala.
}
