% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HumanSurface.R
\docType{data}
\name{HumanSurface}
\alias{HumanSurface}
\title{Human surface markers}
\format{a \code{vector} with surface markers}
\source{
wlabethz - http://wlab.ethz.ch/cspa/
}
\usage{
data(HumanSurface)
}
\description{
vector with names as surface markers and elements as confidence of being a surface marker.
}
\author{
Blanca Pijuan-Sala
}
