% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_reordering.R
\name{leaf_reordering}
\alias{leaf_reordering}
\title{Reorder leaves of hierarchical clustering}
\usage{
leaf_reordering(data, clust.method = "ward.D2", reorder = "column",
  cor.method = "spearman")
}
\arguments{
\item{data}{Matrix of counts with genes as rows and cells as columns}

\item{clust.method}{Clustering method. Default: ward.D2.}

\item{reorder}{What would you like to reorder? Options: "column", "row", "both". Default: "column".}

\item{cor.method}{method of correlation to compute clustering. Default: spearman.}
}
\value{
Resulting list from hierarchical clustering.
}
\description{
It reorders the leaves of the hierarchical clustering so that they 
             depending on their similarities.
}
\author{
Blanca Pijuan-Sala.
}
