% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotGeneLevels.R
\name{plotGeneLevels}
\alias{plotGeneLevels}
\title{Plot gene expression}
\usage{
plotGeneLevels(data, x, y, gene, cols = c("blue", "gold"),
  pointSize = 1, xlab = "x", ylab = "y", titlePlot = gene,
  clusters = NULL, boxplot = FALSE)
}
\arguments{
\item{data}{Dataframe or matrix containing the counts (genes x cells). The cells
should correspond to the ones used for the plot. This matrix is only used to take 
the values of the row of the gene to plot.}

\item{x}{x coordinate. It needs to have entries per each cell in the counts matrix. 
e.g. tsne$Y[,1] when performing tSNE on those cells.The plot does not need to be based
all genes.}

\item{y}{y coordinate. It needs to have entries per each cell in the counts matrix. 
e.g. tsne$Y[,2] when performing tSNE on those cells.The plot does not need to be based
all genes.}

\item{gene}{Gene that you want the expression of. This should be the row ID in 
your data table.}

\item{cols}{Colour range for the expression values. Default: blue,gold.}

\item{pointSize}{cex size for points. Default: 1.}

\item{xlab}{x-label. Default: "x".}

\item{ylab}{y-label. Default: "y".}

\item{titlePlot}{plot title. Default: the gene name introduced in "gene".}

\item{clusters}{If boxplot=TRUE, this should be specified. It should be a vector of
cells as names and colors defining the cluster the cell belongs to. The colours will
be used for colouring the boxplots.}

\item{boxplot}{Logical. TRUE: It will plot a boxplot next to it with the gene 
expression levels per cluster.}
}
\description{
It plots the level of gene expression of a particular gene
 within the plot you want.
}
\note{
The code where boxplot=TRUE is adapted from Ximena Ibarra-Soria.
}
\author{
Blanca Pijuan-Sala.
}
