% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_selectCellsinPlot.R
\name{selectCellsinPlot}
\alias{selectCellsinPlot}
\title{Select cells from a plot}
\usage{
selectCellsinPlot(data, colors = NULL, pch = 20, ...)
}
\arguments{
\item{data}{matrix or dataframe with x and y in columns 1 and 2, respectively.
Ideally, the rownames should correspond to the cell names.}

\item{colors}{row vector with the colors for the data. The indices between
the data and the colors should match. Default: All points will be black.}

\item{pch}{Type of point in plot. Default: 20.}

\item{...}{Other additional features for plot.}
}
\value{
Vector of cells.
}
\description{
It allows you to select cells from a plot
 and it returns a vector of cell names or indices. 
 Select your points by left clicking, and then right clicking to close the 
 polygon.
}
\seealso{
\code{\link[gatepoints]{fhs}}.
}
\author{
Blanca Pijuan Sala.
}
