% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_trainForest.R
\name{trainForest}
\alias{trainForest}
\title{Train randomForest on a dataset.}
\usage{
trainForest(data, groups, numTrees = 1000, rankData = T)
}
\arguments{
\item{data}{Dataframe or matrix containing the counts (genes x cells).
It is recommended to use only the HVG for this.}

\item{groups}{Vector specifying which group each cell belongs to.
Cells should be as names and group ID as vector content.}

\item{numTrees}{Number of trees to grow in randomForest. Default: 1000.
Cells should be as names and group ID as vector content.}

\item{rankData}{Logical. Rank data before training based on counts? 
Default: TRUE}
}
\value{
list with randomForest object (RForest) and a vector of the final
genes used for training.
}
\description{
Cluster cells using the spearman rank correlation
as a similarity metric and (1-correlation)/2 as a distance metric.
}
\references{
\code{\link[randomForest]{randomForest}}.
}
\seealso{
\code{\link[randomForest]{randomForest}},
\code{\link[anSeq]{allocateCellsForest}}
}
\author{
Blanca Pijuan Sala.
}
